# [MINOR] Improve performance of `create_hashes`