# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

`create_hashes` is not super critical to performance but shows up in profiles.
There are some tweaks we can do:

* Don't rehash the first array if we need to hash multiple columns
* Directly iterate over values for primitive / non null arrays 
* Use `value_unchecked`

Overall this is a little bit faster:
```
┃ Query        ┃     main ┃ create_hashes_primitive ┃        Change ┃
┡━━━━━━━━━━━━━━╇━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━┩
│ QQuery 1     │ 189.57ms │                186.99ms │     no change │
│ QQuery 2     │  58.37ms │                 56.41ms │     no change │
│ QQuery 3     │  50.00ms │                 47.19ms │ +1.06x faster │
│ QQuery 4     │  39.27ms │                 37.18ms │ +1.06x faster │
│ QQuery 5     │  95.93ms │                 91.49ms │     no change │
│ QQuery 6     │  10.27ms │                 10.27ms │     no change │
│ QQuery 7     │ 198.99ms │                197.05ms │     no change │
│ QQuery 8     │  73.58ms │                 70.10ms │     no change │
│ QQuery 9     │ 141.44ms │                137.33ms │     no change │
│ QQuery 10    │  97.62ms │                 98.02ms │     no change │
│ QQuery 11    │  45.01ms │                 45.78ms │     no change │
│ QQuery 12    │  67.86ms │                 67.33ms │     no change │
│ QQuery 13    │ 177.16ms │                173.97ms │     no change │
│ QQuery 14    │  12.44ms │                 11.56ms │ +1.08x faster │
│ QQuery 15    │  23.23ms │                 22.73ms │     no change │
│ QQuery 16    │  46.84ms │                 47.63ms │     no change │
│ QQuery 17    │ 680.77ms │                669.99ms │     no change │
│ QQuery 18    │ 521.03ms │                509.00ms │     no change │
│ QQuery 19    │  58.23ms │                 57.57ms │     no change │
│ QQuery 20    │ 204.74ms │                211.60ms │     no change │
│ QQuery 21    │ 263.20ms │                251.50ms │     no change │
│ QQuery 22    │  28.34ms │                 27.72ms │     no change │
└──────────────┴──────────┴─────────────────────────┴───────────────┘
```


# Rationale for this change

<!--
 Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
 Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.  
-->

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are these changes tested?

Covered by existing tests

<!--
We typically require tests for all PRs in order to:
1. Prevent the code from being accidentally broken by subsequent changes
2. Serve as another way to document the expected behavior of the code

If tests are not included in your PR, please explain why (for example, are they covered by existing tests)?
-->

# Are there any user-facing changes?

<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!--
If there are any breaking changes to public APIs, please add the `api change` label.
-->